This commit just finishes what was started on `hack`: an end-to-end, scale-out deployment for both `verifier` and `registrar`.
The fact that, once added to a `verifier`, an `agent` cannot be removed
unless that particular verifier is directly addressed has forced me to
create one service per (`StatefulSet`) `verifier` `Pod`. While still
looking for simpler alternatives, this PR also adds a very crude
implementation of a "client-side consistent hashing" (using the `agent`
UUID), named `/usr/local/bin/kt`

Finally, this PR also adds support for the supplying of custom EK
certificates in a service, required by the `tenant` `Pod`

I consider most of what is here is a "prototype"/"throwaway" code, and
intend to promptly incoporate all this on the `helm` chart.

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>