Deployability and usability improvements
Some minimalistic changes in order to make this helm chart readily
functional, out of the box, by a large contingent of users. The idea is
to allow one to try it out with something simple, such as `minikube`,
with minimal tinkering.

First of all, there is now a very simple set of `curl` commands to grab
`kubectl` in case an image which does not have it is used (I do
understand that not every Kubernetes environment will give pods direct
access to the internet to download, but in this case the deployment
would fail anyway)

Second, the "non-persistent" sqlite deployment the default, and
everything else (either persistent sqlite or other databases) to be
deployed only via a customized set of parameters.

Finally, a have added a warning for non-implemented options, such as
postgres database.

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>