Add a very simple functional testing with `helm-keylime-test" target
This PR adds a new helper script, "kt" which an be used to both "add" or
"delete" ALL registered `agents`, as well as ship files (e.g., boot or
runtime policies) to the `tenant` pod.

For the moment, all that is tested is "identity" attestation, which is
"better than nothing". In subsequent PRs, we will add measured boot and
IMA attestation.

I have also returned the level of privilege on the `agent` pod to
"privileged", as the goal is to have an enviroment which works
"out-of-the-box" on something like Minikube.

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>